GH-438: Move TemporarySpace to *.utils as it is shared
