+++ protobuf-maven-plugin/src/main/java/io/github/ascopes/protobufmavenplugin/dependencies/UrlResourceFetcher.java
@@ -18,8 +18,8 @@ package io.github.ascopes.protobufmavenplugin.dependencies;
 
 import static java.util.Objects.requireNonNullElse;
 
-import io.github.ascopes.protobufmavenplugin.generation.TemporarySpace;
 import io.github.ascopes.protobufmavenplugin.utils.Digests;
+import io.github.ascopes.protobufmavenplugin.utils.TemporarySpace;
 import java.io.BufferedOutputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
