+++ protobuf-maven-plugin/src/main/java/io/github/ascopes/protobufmavenplugin/utils/ConcurrentExecutor.java
@@ -28,7 +28,6 @@ import java.util.stream.Collector;
 import java.util.stream.Collectors;
 import javax.annotation.PreDestroy;
 import javax.inject.Named;
-import javax.inject.Singleton;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
