+++ protobuf-maven-plugin/src/test/java/io/github/ascopes/protobufmavenplugin/utils/TemporarySpaceTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.github.ascopes.protobufmavenplugin.generation;
+package io.github.ascopes.protobufmavenplugin.utils;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
